requirements.txt
aiogram>=3.26,<4
aiohttp>=3.13,<4
Pillow>=12.1,<13
